# apps/api/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "Body_upload_file_sessions__session_id__files_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_file_sessions__session_id__files_post",
        "type": "object"
      },
      "BudgetStatus": {
        "additionalProperties": false,
        "description": "Per-session budget counters and limits.",
        "properties": {
          "file_count": {
            "default": 0,
            "title": "File Count",
            "type": "integer"
          },
          "file_count_limit": {
            "default": 20,
            "title": "File Count Limit",
            "type": "integer"
          },
          "steps_limit": {
            "default": 25,
            "title": "Steps Limit",
            "type": "integer"
          },
          "steps_used": {
            "default": 0,
            "title": "Steps Used",
            "type": "integer"
          },
          "tokens_limit": {
            "default": 150000,
            "title": "Tokens Limit",
            "type": "integer"
          },
          "tokens_used": {
            "default": 0,
            "title": "Tokens Used",
            "type": "integer"
          },
          "tool_calls_limit": {
            "default": 40,
            "title": "Tool Calls Limit",
            "type": "integer"
          },
          "tool_calls_used": {
            "default": 0,
            "title": "Tool Calls Used",
            "type": "integer"
          },
          "wall_seconds_limit": {
            "default": 1500,
            "title": "Wall Seconds Limit",
            "type": "integer"
          },
          "wall_seconds_used": {
            "default": 0,
            "title": "Wall Seconds Used",
            "type": "integer"
          }
        },
        "title": "BudgetStatus",
        "type": "object"
      },
...
      "ErrorCode": {
        "enum": [
          "malformed_csv",
          "unsupported_file_type",
          "file_too_large",
          "upload_limit_exceeded",
          "missing_required_columns",
          "ambiguous_schema",
          "tool_not_registered",
          "validation_loop_exhausted",
          "generated_code_failed",
          "test_failed",
          "command_timeout",
          "repair_cannot_reproduce",
          "author_intent_schema_mismatch",
          "author_custom_workflow_not_validated",
          "author_custom_build_failed",
          "author_validation_failed",
          "universal_validation_failed",
          "contract_specific_validation_failed",
          "generated_pytest_failed",
          "golden_output_comparison_failed",
          "safety_validation_failed",
          "artifact_validation_failed",
          "author_model_required_for_custom_build",
          "author_model_required_for_authoring",
          "author_contract_planning_failed",
          "author_contract_clarification_required",
          "author_contract_review_failed",
          "author_code_generation_failed",
          "author_test_generation_failed",
          "author_generated_code_failed",
          "author_model_did_not_contribute",
          "author_validation_not_contract_driven",
          "author_non_llm_business_logic_detected",
          "author_non_llm_artifact_maker_detected",
          "author_required_artifact_overwritten",
          "budget_exhausted_tokens",
          "budget_exhausted_tool_calls",
          "budget_exhausted_steps",
          "budget_exhausted_wall_time",
          "budget_exhausted_file_count",
          "approval_declined",
          "user_abandoned",
          "sandbox_crash",
          "workflow_interrupted",
          "unknown"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
...
      "FailureMitigation": {
        "additionalProperties": false,
        "description": "Computed guidance for finance users when a session ends in failure.",
        "properties": {
          "can_download_archive": {
            "default": false,
            "title": "Can Download Archive",
            "type": "boolean"
          },
          "can_resume": {
            "default": false,
            "title": "Can Resume",
            "type": "boolean"
          },
          "evidence_items": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Items",
            "type": "array"
          },
          "failure_category": {
            "title": "Failure Category",
            "type": "string"
          },
          "mitigation_actions": {
            "items": {
              "$ref": "#/components/schemas/MitigationAction"
            },
            "title": "Mitigation Actions",
            "type": "array"
          },
          "primary_action": {
            "$ref": "#/components/schemas/MitigationAction"
          },
          "retry_safe": {
            "default": false,
            "title": "Retry Safe",
            "type": "boolean"
          },
          "secondary_actions": {
            "items": {
              "$ref": "#/components/schemas/MitigationAction"
            },
            "title": "Secondary Actions",
            "type": "array"
          },
          "technical_details_ref": {
            "default": "events",
            "title": "Technical Details Ref",
            "type": "string"
          },
          "user_summary": {
            "title": "User Summary",
            "type": "string"
          },
          "user_title": {
            "title": "User Title",
            "type": "string"
          },
          "what_we_found": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "What We Found"
          }
        },
        "required": [
          "user_title",
          "user_summary",
          "primary_action",
          "failure_category"
        ],
        "title": "FailureMitigation",
        "type": "object"
      },
      "FileUploadResponse": {
        "additionalProperties": false,
        "description": "POST /sessions/{id}/files response.",
        "properties": {
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "hash_sha256": {
            "title": "Hash Sha256",
            "type": "string"
          },
          "relative_path": {
            "title": "Relative Path",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "uploaded_at": {
            "format": "date-time",
            "title": "Uploaded At",
            "type": "string"
          },
          "uploaded_file_id": {
            "format": "uuid",
            "title": "Uploaded File Id",
            "type": "string"
          }
        },
        "required": [
          "uploaded_file_id",
          "filename",
          "size_bytes",
          "hash_sha256",
          "relative_path",
          "uploaded_at"
        ],
        "title": "FileUploadResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MitigationAction": {
        "description": "User-facing mitigation actions for failed sessions.",
        "enum": [
          "start_new_session",
          "edit_workflow_description",
          "upload_replacement_file",
          "answer_clarification",
          "retry_same_inputs",
          "download_audit_package",
          "open_technical_details",
          "open_repair",
          "provide_expected_output",
          "simplify_workflow",
          "increase_budget",
          "upload_problem_report"
        ],
        "title": "MitigationAction",
        "type": "string"
      },
...
      "Session": {
        "additionalProperties": false,
        "description": "Full session row.",
        "properties": {
          "budget": {
            "$ref": "#/components/schemas/BudgetStatus"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "current_phase": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Phase"
          },
          "current_step": {
            "default": 0,
            "title": "Current Step",
            "type": "integer"
          },
          "failure_mitigation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FailureMitigation"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Event Id"
          },
          "manifest_schema_version": {
            "default": 1,
            "title": "Manifest Schema Version",
            "type": "integer"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SessionStatus",
            "default": "created"
          },
          "terminal_error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "workflow": {
            "$ref": "#/components/schemas/Workflow"
          },
          "workspace_path": {
            "title": "Workspace Path",
            "type": "string"
          }
        },
        "required": [
          "workflow",
          "started_at",
          "updated_at",
          "workspace_path"
        ],
        "title": "Session",
        "type": "object"
      },
...
      "SessionStatus": {
        "enum": [
          "created",
          "running",
          "paused_user",
          "paused_approval",
          "completed",
          "failed_budget",
          "failed_model",
          "failed_sandbox",
          "failed_user_reject",
          "failed_other",
          "auto_archived"
        ],
        "title": "SessionStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "Workflow": {
        "enum": [
          "author",
          "repair"
        ],
        "title": "Workflow",
        "type": "string"
      },
...
    }
  },
  "info": {
    "description": "Finance-team-facing app for authoring and repairing Python finance agents from Excel/CSV samples. The LLM proposes typed plans and edits; this backend deterministically validates, sandboxes, runs, and records every action.",
    "title": "AgentForge API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/sessions/{session_id}": {
      "delete": {
        "description": "Remove a session from the dashboard (soft) or erase it permanently.",
        "operationId": "delete_session_endpoint_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "description": "When true, permanently removes the session and workspace. When false (default), moves the session to Recently deleted.",
            "in": "query",
            "name": "permanent",
            "required": false,
            "schema": {
              "default": false,
              "description": "When true, permanently removes the session and workspace. When false (default), moves the session to Recently deleted.",
              "title": "Permanent",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Session Endpoint",
        "tags": [
          "sessions"
        ]
      },
      "get": {
        "description": "Fetch session detail (status, phase, budgets, current step).",
        "operationId": "get_session_sessions__session_id__get",
        "parameters": [
          {
            "description": "Session UUID",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "description": "Session UUID",
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "sessions"
        ]
      }
    },
...
    "/sessions/{session_id}/archive.zip": {
      "get": {
        "description": "Stream the full session archive ZIP.\n\nIf ``archive.zip`` does not exist yet (the agent never called\n``archive_workspace``), build it now and emit an\n``ARTIFACT_GENERATED`` event. Subsequent calls return the cached\nfile. Sessions with no load-bearing contents (no reports / outputs\n/ generated / working) yield 409 \u2014 the wizard surfaces this as\n\"nothing to archive yet.\"",
        "operationId": "download_archive_sessions__session_id__archive_zip_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Archive",
        "tags": [
          "files"
        ]
      }
    },
    "/sessions/{session_id}/artifacts/{relative_path}": {
      "get": {
        "description": "Stream an individual artifact from the session workspace.\n\nAllowed prefixes: ``reports/``, ``outputs/``, ``generated/``,\n``working/``. Everything else is rejected at the boundary\n(uploads/ excluded \u2014 they're the user's own files; events.jsonl\n+ manifest.json served via /audit/export and the archive).\n\nINV-5: the path is resolved through ``WorkspaceManager.resolve_in``\nso ``..`` and absolute paths are rejected. Content-Type is\ninferred from the extension (``text/markdown`` for ``.md``,\n``text/csv`` for ``.csv``, etc.).",
        "operationId": "download_artifact_sessions__session_id__artifacts__relative_path__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "relative_path",
            "required": true,
            "schema": {
              "title": "Relative Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "files"
        ]
      }
    },
...
    "/sessions/{session_id}/files": {
      "post": {
        "description": "Upload a CSV or XLSX into the session workspace.\n\nValidates filename extension and MIME type BEFORE reading the body;\nenforces single-file and per-session total size limits; rejects\nuploads if the session has reached its file-count budget.",
        "operationId": "upload_file_sessions__session_id__files_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_file_sessions__session_id__files_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FileUploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload File",
        "tags": [
          "files"
        ]
      }
    },
...
  }
}